requirements.txt
# Core dependencies
PySide6>=6.6.0
qasync>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Audio capture
pyaudio>=0.2.14
numpy>=1.24.0

# STT providers
deepgram-sdk>=3.2.0

# Translation providers
huggingface-hub>=0.20.0

# LLM providers
openai>=1.12.0

# Utilities
websockets>=12.0
aiohttp>=3.9.0
tenacity>=8.2.0
pyyaml>=6.0.1

